# aes: add GCM-TLS checks and improve performance