--- a/openssl/aes.go
+++ b/openssl/aes.go
@@ -460,7 +460,7 @@ func newCipherCtx(cipher C.GO_EVP_CIPHER_PTR, mode C.int, key, iv []byte) (C.GO_
 }
 
 func bigUint64(b []byte) uint64 {
-	_ = b[7]
+	_ = b[7] // bounds check hint to compiler; see go.dev/issue/14808
 	return uint64(b[7]) | uint64(b[6])<<8 | uint64(b[5])<<16 | uint64(b[4])<<24 |
 		uint64(b[3])<<32 | uint64(b[2])<<40 | uint64(b[1])<<48 | uint64(b[0])<<56
 }